Add breadcrumb dynamic component and change prop type location
